scripts/requirements.txt
httpx>=0.27.0
pandas>=2.2.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
supabase>=2.4.0
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
